refactor: Update app.service.ts to use IP geolocation API and weather API
The app.service.ts file has been refactored to use an IP geolocation API and a weather API to retrieve the client's location and temperature. This improves the accuracy of the data and provides a better user experience. The code now makes use of the ipdata.co API for geolocation and the weatherapi.com API for weather information. The client_ip variable has been removed as it is no longer needed. The updated code also includes some minor variable name changes for clarity.